+++ rl_cli/commands/__init__.py
@@ -1,0 +1,4 @@
+"""
+Command modules for rl-cli.
+Each module represents a top-level command group.
+"""
