+++ tests/test_cli.py
@@ -4,6 +4,7 @@ import json
 from unittest.mock import AsyncMock, patch
 import pytest
 from rl_cli.main import run
+from rl_cli.utils import runloop_api_client
 
 @pytest.mark.asyncio
 async def test_devbox_list(capsys):
@@ -25,7 +26,6 @@ async def test_devbox_list(capsys):
     
     # Create mock devboxes
     mock_devboxes = AsyncMock()
-    mock_devboxes.list = AsyncMock()
     
     # Create mock paginator
     async def mock_aiter(self):
@@ -33,10 +33,13 @@ async def test_devbox_list(capsys):
     
     mock_paginator = AsyncMock()
     mock_paginator.__aiter__ = mock_aiter
-    mock_devboxes.list.return_value = mock_paginator
+    mock_devboxes.list = AsyncMock(return_value=mock_paginator)
     mock_api_client.devboxes = mock_devboxes
     
-    with patch('rl_cli.main.runloop_api_client', return_value=mock_api_client), \
+    # Clear the cache to ensure we get a fresh client
+    runloop_api_client.cache_clear()
+    
+    with patch('rl_cli.utils.AsyncRunloop', return_value=mock_api_client), \
          patch('sys.argv', ['rl', 'devbox', 'list']), \
          patch.dict('os.environ', {'RUNLOOP_API_KEY': 'test-api-key', 'RUNLOOP_ENV': 'dev'}):
         await run()
@@ -71,10 +74,18 @@ async def test_blueprint_list(capsys):
     # Create mock blueprints
     mock_blueprints = AsyncMock()
     mock_blueprints.list = AsyncMock()
-    mock_blueprints.list.return_value = AsyncMock(blueprints=[mock_blueprint])
+    
+    # Create mock response
+    class MockResponse:
+        blueprints = [mock_blueprint]
+    
+    mock_blueprints.list.return_value = MockResponse()
     mock_api_client.blueprints = mock_blueprints
 
-    with patch('rl_cli.main.runloop_api_client', return_value=mock_api_client), \
+    # Clear the cache to ensure we get a fresh client
+    runloop_api_client.cache_clear()
+    
+    with patch('rl_cli.utils.AsyncRunloop', return_value=mock_api_client), \
          patch('sys.argv', ['rl', 'blueprint', 'list']), \
          patch.dict('os.environ', {'RUNLOOP_API_KEY': 'test-api-key', 'RUNLOOP_ENV': 'dev'}):
         await run()
@@ -111,7 +122,10 @@ async def test_devbox_get(capsys):
     mock_devboxes.retrieve = AsyncMock(return_value=mock_devbox)
     mock_api_client.devboxes = mock_devboxes
 
-    with patch('rl_cli.main.runloop_api_client', return_value=mock_api_client), \
+    # Clear the cache to ensure we get a fresh client
+    runloop_api_client.cache_clear()
+    
+    with patch('rl_cli.utils.AsyncRunloop', return_value=mock_api_client), \
          patch('sys.argv', ['rl', 'devbox', 'get', '--id', 'test-id']), \
          patch.dict('os.environ', {'RUNLOOP_API_KEY': 'test-api-key', 'RUNLOOP_ENV': 'dev'}):
         await run()
@@ -148,7 +162,10 @@ async def test_devbox_create(capsys):
     mock_devboxes.create = AsyncMock(return_value=mock_devbox)
     mock_api_client.devboxes = mock_devboxes
 
-    with patch('rl_cli.main.runloop_api_client', return_value=mock_api_client), \
+    # Clear the cache to ensure we get a fresh client
+    runloop_api_client.cache_clear()
+    
+    with patch('rl_cli.utils.AsyncRunloop', return_value=mock_api_client), \
          patch('sys.argv', [
             'rl', 'devbox', 'create',
             '--resources', 'SMALL',
@@ -169,6 +186,9 @@ async def test_devbox_create(capsys):
 @pytest.mark.asyncio
 async def test_missing_api_key():
     """Test error handling when API key is missing."""
+    # Clear the cache to ensure we get a fresh client
+    runloop_api_client.cache_clear()
+    
     with patch.dict('os.environ', {}, clear=True), \
          patch('sys.argv', ['rl', 'devbox', 'list']), \
          pytest.raises(RuntimeError) as exc_info:
